+++ generators/client/generator.ts
@@ -23,6 +23,7 @@ import BaseApplicationGenerator from '../base-application/index.ts';
 import { createNeedleCallback } from '../base-core/support/index.ts';
 import { LOGIN_REGEX_JS } from '../generator-constants.js';
 import { GENERATOR_CLIENT, GENERATOR_COMMON, GENERATOR_CYPRESS } from '../generator-list.ts';
+import { isReservedTypescriptKeyword } from '../javascript/support/reserved-words.ts';
 
 import { addEnumerationFiles } from './entity-files.ts';
 import { writeFiles as writeCommonFiles } from './files-common.ts';
@@ -194,7 +195,50 @@ export default class ClientGenerator extends ClientApplicationGenerator {
     return this.delegateTasksToBlueprint(() => this.preparing);
   }
 
-  // Public API method used by the getter and also by Blueprints
+  get preparingEachEntity() {
+    return this.asPreparingEachEntityTaskGroup({
+      preparing({ entityName }) {
+        if (isReservedTypescriptKeyword(entityName)) {
+          throw new Error(`The entity name "${entityName}" is a reserved TypeScript keyword. It may cause issues in your application.`);
+        }
+      },
+    });
+  }
+
+  get [ClientApplicationGenerator.PREPARING_EACH_ENTITY]() {
+    return this.delegateTasksToBlueprint(() => this.preparingEachEntity);
+  }
+
+  get preparingEachEntityField() {
+    return this.asPreparingEachEntityFieldTaskGroup({
+      preparing({ entity, field }) {
+        if (isReservedTypescriptKeyword(field.fieldName)) {
+          throw new Error(`The field name "${field.fieldName}" in entity "${entity.name}" is a reserved TypeScript keyword.`);
+        }
+      },
+    });
+  }
+
+  get [ClientApplicationGenerator.PREPARING_EACH_ENTITY_FIELD]() {
+    return this.delegateTasksToBlueprint(() => this.preparingEachEntityField);
+  }
+
+  get preparingEachEntityRelationship() {
+    return this.asPreparingEachEntityRelationshipTaskGroup({
+      preparing({ entity, relationship }) {
+        if (isReservedTypescriptKeyword(relationship.relationshipName)) {
+          throw new Error(
+            `The relationship name "${relationship.relationshipName}" in entity "${entity.name}" is a reserved TypeScript keyword.`,
+          );
+        }
+      },
+    });
+  }
+
+  get [ClientApplicationGenerator.PREPARING_EACH_ENTITY_RELATIONSHIP]() {
+    return this.delegateTasksToBlueprint(() => this.preparingEachEntityRelationship);
+  }
+
   get writing() {
     return this.asWritingTaskGroup({
       async cleanup({ application, control }) {
