+++ generators/javascript/generator.ts
@@ -18,15 +18,11 @@
  */
 
 import { packageJson } from '../../lib/index.ts';
-import BaseApplicationGenerator from '../base-application/index.ts';
 import BaseSimpleApplicationGenerator from '../base-simple-application/index.ts';
 
-import { isReservedTypescriptKeyword } from './support/reserved-words.ts';
 import type {
   Application as JavascriptApplication,
   Config as JavascriptConfig,
-  Entity as JavascriptEntity,
-  JavascriptSimpleApplication,
   Options as JavascriptOptions,
   Source as JavascriptSource,
 } from './types.ts';
@@ -35,24 +31,13 @@ import type {
  * Utility class with types.
  */
 export class JavascriptSimpleApplicationGenerator extends BaseSimpleApplicationGenerator<
-  JavascriptSimpleApplication,
-  JavascriptConfig,
-  JavascriptOptions,
-  JavascriptSource
-> {}
-
-/**
- * Utility class with types.
- */
-export class JavascriptApplicationGenerator extends BaseApplicationGenerator<
-  JavascriptEntity,
   JavascriptApplication,
   JavascriptConfig,
   JavascriptOptions,
   JavascriptSource
 > {}
 
-export default class JavascriptGenerator extends JavascriptApplicationGenerator {
+export default class JavascriptGenerator extends JavascriptSimpleApplicationGenerator {
   async beforeQueue() {
     if (!this.fromBlueprint) {
       await this.composeWithBlueprints();
@@ -61,50 +46,6 @@ export default class JavascriptGenerator extends JavascriptApplicationGenerator 
     if (!this.delegateToBlueprint) {
       await this.dependsOnBootstrap('javascript');
     }
-  }
-
-  get preparingEachEntity() {
-    return this.asPreparingEachEntityTaskGroup({
-      preparing({ entityName }) {
-        if (isReservedTypescriptKeyword(entityName)) {
-          throw new Error(`The entity name "${entityName}" is a reserved TypeScript keyword. It may cause issues in your application.`);
-        }
-      },
-    });
-  }
-
-  get [JavascriptApplicationGenerator.PREPARING_EACH_ENTITY]() {
-    return this.delegateTasksToBlueprint(() => this.preparingEachEntity);
-  }
-
-  get preparingEachEntityField() {
-    return this.asPreparingEachEntityFieldTaskGroup({
-      preparing({ entity, field }) {
-        if (isReservedTypescriptKeyword(field.fieldName)) {
-          throw new Error(`The field name "${field.fieldName}" in entity "${entity.name}" is a reserved TypeScript keyword.`);
-        }
-      },
-    });
-  }
-
-  get [JavascriptApplicationGenerator.PREPARING_EACH_ENTITY_FIELD]() {
-    return this.delegateTasksToBlueprint(() => this.preparingEachEntityField);
-  }
-
-  get preparingEachEntityRelationship() {
-    return this.asPreparingEachEntityRelationshipTaskGroup({
-      preparing({ entity, relationship }) {
-        if (isReservedTypescriptKeyword(relationship.relationshipName)) {
-          throw new Error(
-            `The relationship name "${relationship.relationshipName}" in entity "${entity.name}" is a reserved TypeScript keyword.`,
-          );
-        }
-      },
-    });
-  }
-
-  get [JavascriptApplicationGenerator.PREPARING_EACH_ENTITY_RELATIONSHIP]() {
-    return this.delegateTasksToBlueprint(() => this.preparingEachEntityRelationship);
   }
 
   get writing() {
@@ -118,7 +59,7 @@ export default class JavascriptGenerator extends JavascriptApplicationGenerator 
     });
   }
 
-  get [JavascriptApplicationGenerator.WRITING]() {
+  get [JavascriptSimpleApplicationGenerator.WRITING]() {
     return this.delegateTasksToBlueprint(() => this.writing);
   }
 
@@ -162,7 +103,7 @@ export default class JavascriptGenerator extends JavascriptApplicationGenerator 
     });
   }
 
-  get [JavascriptApplicationGenerator.POST_WRITING]() {
+  get [JavascriptSimpleApplicationGenerator.POST_WRITING]() {
     return this.delegateTasksToBlueprint(() => this.postWriting);
   }
 }
