+++ generators/spring-boot/generator.ts
@@ -34,7 +34,7 @@ import { mutateData } from '../../lib/utils/index.ts';
 import BaseApplicationGenerator from '../base-application/index.ts';
 import { createNeedleCallback, isWin32 } from '../base-core/support/index.ts';
 import { editPropertiesFileCallback } from '../base-core/support/properties-file.ts';
-import type { Config as ClientConfig } from '../client/types.ts';
+import type { Config as ClientConfig, Entity as ClientEntity } from '../client/types.ts';
 import type { Source as CommonSource } from '../common/types.ts';
 import type { Entity as CypressEntity } from '../cypress/types.ts';
 import { ADD_SPRING_MILESTONE_REPOSITORY } from '../generator-constants.js';
@@ -447,7 +447,7 @@ ${classProperties
     return this.asPreparingEachEntityTaskGroup({
       prepareEntity({ entity, application }) {
         if (entity.entityRestLayer === false) {
-          entity.entityClientModelOnly = true;
+          (entity as unknown as ClientEntity).entityClientModelOnly = true;
         }
 
         const hasAnyAuthority = (authorities: string[]): string | undefined =>
@@ -576,7 +576,7 @@ ${classProperties
           entityJavaCustomFilters: sortedUniqBy(entity.fields.map(field => field.propertyJavaCustomFilter).filter(Boolean), 'type'),
         });
 
-        mutateData(entity as CypressEntity, {
+        mutateData(entity as unknown as CypressEntity, {
           __override__: true,
           // Reactive with some r2dbc databases doesn't allow insertion without data.
           workaroundEntityCannotBeEmpty: ({ reactive, prodDatabaseType }: any) =>
